student_engagement_monitor/requirements.txt
# Core Computer Vision & ML Libraries
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
mediapipe==0.10.8
xgboost==2.0.3
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4

# Deep Learning (for advanced features)
torch==2.1.2
torchvision==0.16.2
ultralytics==8.1.0  # YOLOv8 for person/object detection

# Video Processing
imageio==2.33.1
imageio-ffmpeg==0.4.9

# Dashboard & Visualization
streamlit==1.29.0
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1
pillow==10.1.0

# Data Download
kaggle==1.5.16

# Tracking & Detection
filterpy==1.4.5  # Kalman filter for tracking
scipy==1.11.4

# Report Generation
reportlab==4.0.7
fpdf2==2.7.6

# Optional: GPU acceleration
# onnxruntime-gpu==1.16.3  # Uncomment if using ONNX models with GPU

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
